fixed: int->float in closeness-centrality

+++ cpp_easygraph/functions/centrality/closeness.cpp
@@ -8,7 +8,7 @@ double closeness_dijkstra(const Graph_L& G_l, const int &S, int cutoff, Segment_
     const int dis_inf = 0x3f3f3f3f;
     int N = G_l.n;
     segment_tree_zkw.init(N);
-    std::vector<int> dis(N+1, INT_MAX);
+    std::vector<float> dis(N+1, INT_MAX);
     const std::vector<LinkEdge>& E = G_l.edges;
     const std::vector<int>& head = G_l.head;
     int number_connected = 0;
